agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "livekit>=1.0.23",
    "livekit-agents[openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-silero>=0.2.0",
    "livekit-plugins-turn-detector>=1.3.12",
    "python-dotenv>=1.2.1",
]
